[flow][match] Add custom errors for invalid match case syntax
Summary:
Add parsing and custom errors for invalid match case syntax that may commonly occur.

Added this as a Flow error rather than parser error, so we have an opportunity to add quick-fixes later.

Since it may be a common to change `switch` to `match` on existing code, coalesce multiple of these errors to a single error so we don't spam the user with a million errors.

Changelog: [internal]

In the future, I want to take a look at how we can dedupe some of the parsing between match expressions and match statements.

Reviewed By: SamChou19815

Differential Revision: D76792239

fbshipit-source-id: c68f6d7f9c782cbddcdcaf2bf2e988e12615439c