[flow] Best effort codemod to fix natural inference errors by inserting annotation on object literal declaration
Summary:
This diff implements a codemod to add annotations to reduce errors for the next phase of natural inference: object literals.

It will first find all `const obj = {...}`, and record the tvar id and location of the definition id. We then track repositions of these tvars so that when a repositioned reference flow into another type, we can record it as the upper bound for the object literal definition. This is helpful to find a suitable annotation the following common case:

```
const obj = {...}

f(obj) // where f: (obj: ObjAlias) => mixed
```

In the codemod, we look through all the unsuppressed errors where the error locations are at reference locations of these object literal declaration. We then use the bound we tracked during checking to annotate the object. For now, I keep the max_size=1, so that we are only inserting a single type alias type.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D76829438

fbshipit-source-id: c50b3aadab385a60dff40892c60ef24a11432e86