[flow][match] Add quickfix for not exhaustive match error
Summary:
Since I started working on this project, I had the vision of the user writing an empty `match` and then clicking a button to fill in all the cases. With this diff that vision is now possible!

1) Add `to_ast` functions to the various pattern types in `exhasutive.ml`
2) Store these pattern asts in the error
3) For both `match` statements and expressions
  a. Expression: add the missing cases with an `undefined` body
  b. Statement: add the missing cases with an empty block `{}` body

Note that ultimately we use heuristics to calculate the missing patterns in the exhaustive checking algorithm, so the user may need to make some fixes/add imports as required.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D77125735

fbshipit-source-id: 59a9d68d14bc971e75995cdc411a8b7c38b9bbb5