[flow][constant-condition] A bug fix on the evaluated value of Unary Not
Summary:
There's a bug that we don't flip the boolean value of the argument of Unary Not (`!`). This bug will become more apparent after we launch the evaluation on `true` and `false` literals for milestone 2. I tried to refactor the recursion to not use the `should_report_error_ref` flag but I don't see an easy way out (unless we don't pick where we error and just error on the outer level).

It's unclear to me if erroring on the entire `test` expression (the outmost layer) is easier for IDE quick fixes (maybe?). There will be more clarity as we work down the line.

Changelog: [errors]

Reviewed By: SamChou19815

Differential Revision: D77175054

fbshipit-source-id: c369751a048518d7aa261410f54d9a3c8a502b84