[flow][constant-condition] Add warning option to error message
Summary:
There should be no behavior changes. This diff adds the warning option for constant condition error message. For errors generated from AST side, we can know for sure the error is valid, but for errors generated from type inference, there's a chance that the error is not accurate.

Changelog:[internal]

Reviewed By: SamChou19815

Differential Revision: D76849973

fbshipit-source-id: 42f15ccb383ac1146023bc2cb42d30c554a3a084