[flow][errors] Add error for truthy strings
Summary:
Some strings are truthy and the information is most likely from refinement. Turning on a small amount of errors to show the warning in error messages. We plan to show the warning message to all errors generated from type inference due to possible bugs in the inferrence prodcess. **This change will not be gated by a flag due to the limited amount of errors (100s).**

Changelog: [errors]

Reviewed By: panagosg7

Differential Revision: D77186849

fbshipit-source-id: 99d29cc60a1642b755e3e311e94fc319f92fe277