[flow][match] Add tests for refinements used by bindings
Summary:
We figure out the type of bindings using the refinements system.
We used to have many tests for this (`matching.js`), since the old exhaustive checking system used the refinement system.
But, we moved to the new (more powerful) exhaustive checking system that doesn't use the refinement system, so all those tests were updated.
We still need to test the output of the refinement system however, so people have accurate results for their bindings.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D77413104

fbshipit-source-id: 22ef3a8767ff6f7d2c7d65996b9ec61454499c78