[flow] Opaque type with lower bound 2/n: typing
Summary:
In this diff, I implemented type checking logic for opaque type with lower bound. It's very straightforward, since we just need to add a handler for `lower ~> OpaqueT {lower_t; _}`. A lot of the changes are for renaming, because currently the upper bound is recorded in `super_t`, which might confuse us because the syntax to specify lower bound is `super Lower`.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D77681747

fbshipit-source-id: d18947031c06dc5ab8b0c567ef35a69086c199f0