[flow] Opaque type with lower bound 3/n: gate the new syntax with cosmetic error
Summary:
Since the previous two changes are very straightforward, we don't really need to gate it for further development. We only need to gate so that people don't introduce this before the tooling is ready. Therefore, I decide to gate it with a cosmetic error, which doesn't affect the underlying support at all.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D77685034

fbshipit-source-id: d67ca9b8bef111b9c941829f8169e231dbc03e40