[flow][lsp][test] Support pull diagnostics for tests
Summary:
I'm frustrated with the random failure of `lsp/detailed_diagnostics` test. It's a result of non-deterministic timing of `textDocument/publishDiagnostics` notification. To resolve the issue, I propose that we implement pull diagnostics support, but don't declare corresponding server capability. In this way, it will have no effect on end users, but we now have a way to deterministically pull errors for testing.

Why is it deterministic? Well, because the request will be in the same ordered queue of events we send to the server, and we can always expect a response from the server after we send the request.

This diff implements the behavior on the server side. It aims to match current behavior differences wrt opened files vs all other files: opened files' errors are more decorated (detailed diagnostics, refinement highlighting, etc).

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D79177591

fbshipit-source-id: ff88ba4e51aa2682854a4d3b9e4849f14b831fb7