[flow][strict-comparison] Add `StrictComparison` to `encl_ctx` when reconstructing AST for strict equals
Summary:
When verifying whether the new strict comparison check achieves one-to-one mapping with existing checks, this snippet wasn't flagged by the new check. (https://www.internalfb.com/diff/D78928698?dst_version_fbid=1085627823669809&transaction_fbid=1079712727014366)

```
{
  type T = 'foo' | 'bar';
  declare const x: ?T;
  x === 'foox';
}
```

This is because when reconstructing AST for the strict comparison, `'foox'` is considered as a string instead of a literal. In this diff, we update the enclosing context so that `foox` is considered a string literal to unblock strict comparison.

Reviewed By: gkz

Differential Revision: D79148291

fbshipit-source-id: 93e7c8679bf89c63c53e7ff4c20e36a09a0765b4