[flow][strict-comp] Add detailed error message for `empty` errors
Summary:
This diff adds detailed help message for `===empty` like errors to better help engineers understand flow result.

It's a good question to ask that whether we want to error on empty values, because https://flow.org/en/docs/types/empty/ says "all operations are permitted on something that has the empty type". It just doesn't make too much sense to me to do anything with `empty`. It feels like a user error that it should be fixed, so we error here for strict-comparison.

Additionally, that flow doc is a bit old and there's a general trend to error on the strict side for modern languages.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D79598070

fbshipit-source-id: 097b5df2408f6612a355dde5374d37bdbffbdb7a