[flow] Dump types for tool consumption
Summary:
This diff exposes a new API to dump-types that accesses a JSON representation of Type.t, rather than the existing normalized Ty.t. The idea is that this richer type information can be parsed by external tools which may need the full understanding of Flow's types.

We basically emit in JSON everything that is included in a Type.t except reason and location information. When we see an OpenT we try to resolve it and emit it, but we have a depth counter to ensure that we don't diverge.

For now this API is undocumented.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D76273439

fbshipit-source-id: 68cb0ba72113fea4646d402080f3b4d576566bae