[flow][refactor] Split `LookupProp` into `LookupPropForTvarPopulation` and `LookupPropForSubtyping`
Summary:
Most of the `LookupProp` action are used for subtyping, which can result in `EPropNotFound` errors that are actually subtyping errors. We want to make these errors eventually `incompatible-type` errors, so we need to clearly distinguish them first.

I would like to just renamed it to be `LookupPropForSubtyping`, but there are two `LookupProp` in `react_kit` seem to used to do something like intrinsic indexed access. Therefore, I will split it for now.

In later diffs, I will extend `LookupPropForSubtyping` with more information, so that it can be used to easily generate an `EPropNotFoundInSubtyping` error.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D79752813

fbshipit-source-id: 239bca49dd8282fdec44320d80a431f4c303d546