[flow][error-code-update] Implement new behavior
Summary:
This diff implements the new behavior proposed in the previous diff.

The error code change is quite straightforward. For now, I will normalize `incompatible-call`, `incompatible-cast`, `incompatible-extend`, `incompatible-return`, `incompatible-type-arg`, subtyping related `prop-missing` to `incompatible-type`

For suppression calculation, I decide to go for a simplistic approach with `compatibility` mode: allow all of the error code to be suppressed with `incompatible-type`. Note that this will make some of the non-subtyping related errors be suppressed with `incompatible-type` as well, but it's probably fine since the `compatibility` mode is supposed to be temporary while we mass update suppressions.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D79766489

fbshipit-source-id: 49975fd4289e2df9f0ab34b299fe6c7ae793b7d0