[flow][error-code-update] Introduce flowconfig to gate error code update
Summary:
We should change all the `prop-missing` errors due to subtyping and all other variants like `incompatible-call`, `incompatible-return` to have `incompatible-type` errors for our own sanity, and makes experimenting with error message easier.

This will be a tough migration that will be hard to rollout, so I decide to gate it by config. The config will have 3 variants:

- `old`: status quo
- `new`: all normalized to `incompatible-type`
- `compatibility`: still show the old code, but allow them to be suppressed with `incompatible-type`. This will be the key to rollout in large codebases with active changes

This diff first introduces the config and pass it to various places. The actual implementation of the new behavior will be done in the next diff.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D79766487

fbshipit-source-id: 4e64d02423cfd8303ca0d580a1841f5f8fb01f50