[flow] Add supported_operating_systems flag in flowconfig to restrict to CentOS
Summary:
Adds the ability to restrict a flow server from running unless the OS matches the allowlist.

To distinguish between different linux flavors we look at `/etc/os-release`. If for any reason that step fails we'll return `None` and refuse to start the server. I considered being less conservative and defaulting to start if our detection logic does not work, but decided that it was better to be stricter. If the detection logic blocks legit use cases we can disable the flag and fix it in a future release.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D79818105

fbshipit-source-id: 24894396bf361499a91a2a6e26a31fd2025a79fa