[flow][error] Add line breaks between main error message and explanations
Summary:
For non-speculation errors, this diff now starts to print errors' explanations starting on a new line, and add another line break between the main error message and the explanation. For now, the behavior for speculation errors are unchanged.

It makes explanations more readable, which prepares for D77034460 that will add more explanations for invariant subtyping errors.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D80144584

fbshipit-source-id: 83bc37991acf8faf5546d3e7e0b936ddee1149f9