[flow] experimental type-of-name-experimental command
Summary:
Experimental CLI command that takes the name of a FILE and a NAME and returns
1. the type of the first instance of NAME as identifier that Flow finds in FILE,
2. if (1) fails, the type of a variable found in the export index that most closely matches NAME,
3. if (2) fails too, then returns an appropriate error.

Implementation for (1) is straightforward. For (2) we construct fake contents of the form
```
/* flow */ import NAME from 'SOURCE';
```
where SOURCE is the filename that the query of export index returned. We check these contents on the fly as file FILE, and repeat the process for (1) on NAME.

Depending on the kind of export we adjust the import. For builtins we just include the variable or the type as body.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D80370737

fbshipit-source-id: d86616623d91f81bebd6e97c793071fccfbcc5df