[flow] Error on codeless suppressions
Summary:
This diff unbreaks the codeless suppression detection. Previously, it was a warning, but even the warning was lost under `experimental.require_suppression_with_error_code`. This diff fixes it by treating codeless suppressions more like `EMalformedCode`: we raise it as an error as we are parsing the suppression comments.

The errors will be unsuppressable, but it's fine since you will have an unsuppressed error anyways. By fixing the suppression, you will fix both. The only complication is that it will error on unused codeless suppressions, but it's probably a good idea to clean them up anyways.

Changelog: [errors] Suppressions without error code will now be an error instead of a warning.

Reviewed By: panagosg7

Differential Revision: D81259570

fbshipit-source-id: 383645b9df602ef565295823abd021eec0a1d3e1