[flow][type-guards] record this-type guard on regular class method despite consistency error
Summary:
This diff fixes an inconsistency in how we handle this-type guards on methods of regular functions.

As a reminder, the following definition is an error
```
export class C {
    isA(): this is A { // error
        return this instanceof A;
    }
}
export class A extends C {}
```
because we don't perform consistency checking on `this` type guards.

The inconsistency was that within the same file, the following was still allowed, since the type guard was recorded
```
declare var c: C;
if (c.isA()) {
    c as A;
}
```
but if this code was in another file, the type guard was dropped.

This diff fixes this by changing type_sig_merge to record the type guard in this case.

Changelog: [fix] Fixes an inconsistency in recording the this-type guard on a regular (non-`declare`) method in the same vs different files.

Reviewed By: SamChou19815, NicolasRoelandt

Differential Revision: D81591701

fbshipit-source-id: 2a6360a4caa6fdaa11d79647222da82848322155