[flow][enums] Add lint for requiring explicit enum patterns in match, when enum has unknown members
Summary:
Similar to the existing `require-explicit-enum-switch-cases` lint error, expect for `match`.
One difference is that we only trigger the error when there was a Flow Enum with unknown members (as otherwise, you can simply omit the `default`).

Changelog: [feature] Added a Flow lint `require-explicit-enum-checks` which works like `require-explicit-enum-switch-cases` but for `match`.

Reviewed By: SamChou19815

Differential Revision: D81518883

fbshipit-source-id: fd29b1c72c20438bd23c079cad7a53eddf06569b