[flow] Create a flag to gate natural inference for jsx
Summary:
We want to have natural inference for jsx literals to stop inferring a super specific type for jsx elements in most of the cases. We want this, because these super specific type references the deprecated `ExactReactElement_DEPRECATED`, and it can frequently cause constrain write errors while most of the user don't care about the specific type.

D81724272 shows that it can happen with low enough error diffs, so this diff sets up the flags first.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D81784227

fbshipit-source-id: 1f20ae505ec13412e9d2a21bb487a782da66e30b