[flow] Experimental implementation of natural inference for jsx (gated)
Summary:
This diff implements the initial cut of natural inference for jsx. It's still buggy and not production ready (as shown by some weird errors in D81785156), but it will allow us to play with it in product code under the flag.

The core idea is: if a jsx literal doesn't have hint, it will use a generalized type. For the generalized type, we will try to generalize it to a render type first. If it fails, we will fall back to `React.MixedElement`.

Note that we cannot make use of `return_hint` that's already available, because [during implicit instantiation it will be set to unavailable](https://github.com/facebook/flow/blob/671d8577db2d8748fdb3e805983d83f697ff42ec/src/typing/implicit_instantiation.ml#L961). Therefore, we have to pass in another `should_generalize` (only accounts for hint right now, could also consider `as const` in the future.). During implicit instantiation, `should_generalize` is always false, since it's not helpful to constrain targs with a generalized type that loses type information.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D81785157

fbshipit-source-id: 4c6a9cd4f08fb6ef7633b82c87b657b83802743c