[flow] Combine object props invariant subtyping errors
Summary:
In this diff, we will start to combine object prop invariant subtyping errors. When we are matching all props from LHS and RHS, we will look for the following conditions:

1. Usual conditions where we do invariant subtyping error improvements (no speculation, all resolved)
2. Prop with the given name exists on both sides
3. Both sides have neutral polarity

Notably, 2 and 3 represent the most common case of object subtyping, when the user is not consciously making the upper bound readonly. It's the most likely place to trigger invariant subtyping errors that will otherwise pass covariantly. When there is an error, there is a good chance that there will be multiple of them.

Currently, we will generate a detailed error with similar fix suggestions for every one of them. It significantly increases the output and drowns the user with information. Instead, we can group them together into a single one, like

```
Cannot bla because properties `p` and `q` of object literal [1] are not exactly the same as
those of `B` [2].

The above-mentioned two types must be the same because properties `p` and `q` are invariantly typed. To fix the error,
- Either annotate object literal [1] with `B` [2]
- Or make properties `p` and `q` in object type [2] readonly.
```

For now, I make the case of single property invariant subtyping error still stick to the current behavior (also print out the type of two incompatible properties).

Changelog: [errors] Multiple object property invariant subtyping errors on the same object will be combined into one error. Some error locations might be moved.

Reviewed By: panagosg7

Differential Revision: D81689606

fbshipit-source-id: 65d02b6bade2cb13e5f72f5915c02ccf9bb215d7