[flow] Track whether `React.GetConfig` is coming from userland `React.ElementConfig`
Summary:
D82271511 describes how the new mode for react-ref-as-prop can be hard to rollout.

What's especially difficult is point 3 in the diff summary:

```
declare component A(ref?: React.RefSetter<I>);
component B(...props: React.PropsOf<A>) {...}
```

will fail, because now the props include the ref prop, which will fail the no-ref-in-spread check.

To make it slightly easier to rollout, I plan to implement an intermediate mode where `React.ElementConfig` doesn't include the ref prop, so that we can keep the existing behavior to make cases like the above example still pass, and clean it up later.

This diff doesn't implement the behavior yet. It simply propagates the information all the way to react_kit.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D82568517

fbshipit-source-id: ff6017d48815c4ab40062dbbd90a6366f543a5f8