[flow] Restrict typing of searchElement in array.includes and friends
Summary:
This diff restricts the type of values passed into array .includes, .indexOf, and .lastIndexOf method calls to be valid members of the arrays themselves. This change detects many cases where there was probably a mistake by the programmer (e.g. testing if an array of IDs contains an object with a specific ID, rather than if it contains the ID itself).

 There are still many errors that are likely innocuous in practice, such as cases where we infer arrays like `export const arr = ['FOO', 'BAR']` to have type `Array<'FOO' | 'BAR'>`. It is now an error to do `arr.includes('BAZ')`, which seems incorrect to me because `arr` was not annotated as being limited to only those types.

Changelog: [errors] Flow now strictly enforces that Array.includes, .indexOf, and .lastIndexOf take only values that could potentially be present in the array -- it is now an error to check if arrays include values whose types are unrelated to the array's element type. This catches many cases where developers accidentally test the wrong thing, which may lead to errors at runtime.

Reviewed By: benmonro, SamChou19815

Differential Revision: D82476988

fbshipit-source-id: de6ff9b01c6c4d076b59aa80e391b87d37502ace