[flow][RFC] Just allow `keyof`
Summary:
The genie is out of the bottle. We already have to use it to create homomorphic mapped types. As we are planning to converge with TS syntax, we should just allow it everywhere.

For now, I gated it together with `ts_utility_syntax`, which we intend to flip soon

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D82779207

fbshipit-source-id: 51385ce7b40935e9c8194781996b1fcf76cd16f6