[flow] Produce a hint level diagnostic at the switch stmt eligible for match conversion
Summary: This diff will produce a hint level diagnostic at the switch stmt eligible for match conversion. This is not very visually distracting, but in case the user notices it (most likely to happen when they are writing it for the first time), they can also notice the quickfix option when they hover on it, and hopefully chooses to run it.

Reviewed By: gkz

Differential Revision: D82991447

fbshipit-source-id: eff7a2d2972e370331c3f2eddbc8edea0cf454c6