[flow] Add support for Values utility type
Summary:
For completeness, I added error message for `Values` when ts_syntax is not enabled. It might be slightly confusing because there's no `Values` in TS. This confusion is fine because we show the replacement type in the error message and we don't planned to turn off `ts_utility_syntax`, so it won't really error.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D83017912

fbshipit-source-id: 2a3e027ec8bcffe1e389a92ae8f1a3606bbd8d7c