[flow] Use def_loc in explanations for object invariant subtyping
Summary:
I finally found out why sometimes the quickfix is adding annotations on the second `foo` in

```
const foo = {...}
expect(foo)
```

Turns out we are not putting def_loc in explanations, which is deciding the insertion location of the quickfix. This diff switches it to use def_loc.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D83367829

fbshipit-source-id: 1257003b464be4dbd993bdb930a12221ebaf0849