[flow] No longer use `EvalT` infra to make things hook like in env_resolution
Summary:
My goal is to gradually simplify `MakeHooklikeT` so that it no longer needs to be an `EvalT` in the future.

In env_resolution, we are working with fully resolved types, and we are already under an `OpenT`. Therefore, we can just use a fully resolved `OpenT` with that's lazily populated with result of flowing the input to `HooklikeT`, instead of going through the `EvalT` infra.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D83842078

fbshipit-source-id: 2e0200fba32d98c81de1b991ea5b2a6a410d426a