[flow][errors] Fix ranking with nested speculation errors
Summary:
We have a complex mechanism to score branches in speculation error, and only show those that have high scores. While we have heuristics for specific use_op frames, it generally follows the principle of "longest-frame-wins" strategy.

i.e. errors with `Frame(Prop, Frame(Prop, Frame (Prop, root))))` should have higher score than (`Frame (Prop, root)` , since it's more specific and is more likely to be the final small issue that the user needs to fix.

Unfortunately, we have a bug wrt to the mechanism in nested speculation. In nested speculation, we will wrap the common parent in `Op (Speculation use_op)`, with the goal of not showing the same frames in each branch of speculation over and over again. However, the wrapped common part is not considered for scoring, which can leads to branches with the same score even if they shouldn't. Consider:

```
- Frame(Prop("a"), root)
  - Frame(Prop("inner1"), Op (Speculation (Frame(Prop("a"), root)))) -> case 1
  - Frame(Prop("inner2"), Op (Speculation (Frame(Prop("a"), root)))) -> case 2
- Frame(Prop("b"), root) -> case 3
```

The current scoring mechanism will give all 3 cases the same score, while intuitively case 1 and case 2 should have higher scores.

`React.RefSetter<...>` subtyping errors are affected by this bug, since it often requires nested speculation (the outer union type + the inner union type between instance and null)

This diff fixes the issue by also computing the score for the use_op in `Op (Speculation ..)`. I noticed the issue when I'm refactoring the error generation. For so long, the ranking bug is hiding in plain sight...

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D84103703

fbshipit-source-id: 7274f9b8365fc6f72dee2aa614ef49f0210d803c