[flow][RFC] Simplify roots of error messages 1/n: merge root and explanations where it makes sense
Summary:
The root part of the message is usually something like "Cannot call ... because". In many cases, I believe that these messages are just not useful. Consider

```
foo(bad_arg)
    ^^^^^^^

// cannot call `foo` with `bad_arg` because ...
```

The message is repeating what the user can directly see. For CLI error, the error will include snippet at the error location. For IDE errors, the user can directly see the error site to figure out what's the operation that contains the error.

Hence, I propose that we should get rid of most of the root messages. However, some of the roots are still useful: they serve as short explanations. This diff starts from a few root that also have explanations, and shift the explanation to the root.

Technically, it is implemented through a new variant of root: `ShortExplanationRoot` (while all the legacy ones are called `OperationRoot`). `ShortExplanationRoot` will be rendered like:

```
The `ref` parameter must be a subtype of `React.RefSetter`: <detailed incompatible errors>
```

Later, we should convert all the useless roots like `Cannot assign a to be because <detail>` to be something like: `Type incompatibility: <detail>`.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D83694099

fbshipit-source-id: dfd103955a0374a4380fabc4b242efcdc1e0b07c