[flow][errors] Put frame first in intuitive order for all errors
Summary:
Building on top of the previous diff, this diff changes the style of framed errors to align with stacked errors for all errors.

i.e.

Instead of

```
... because A is incompatible with B in property foo of type argument T
```

We now do

```
... because in type argument T > in property foo: A is incompatible with B
```

This is much more understandable, and the inline references are in the same order of error hierarchy.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D84853811

fbshipit-source-id: 47c074834d615c9d1c84528c3fa7186684833ffb