[flow][RFC] Stacked subtyping errors
Summary:
This diff implements stacked subtyping error formatting, similar to the [formatting in TS](https://www.typescriptlang.org/play/?#code/C4TwDgpgBAQghgJygXigbQN4CM4C8BcUAgggnCADwDOwCAlgHYDmAfAL4C6A3AFCiRQAYgHthKKNkSF4CNrx4ATCAGMANomgAzAK4NlwOsIZRh25QAsAFJtGERwgJSEAbsLoLeyozSg2xqSQRCTBwCdABmDk45HlMLa1EHXiA).

The stack is obtained from the stack of frames. Unlike TS, this implementation will only show stacks where the upper is a named type. If the upper is not a named type, then the frame won't show the incompatibility at this level

e.g. instead of

```
- <desc> is incompatible with <non-named desc> in property `foo`
  - <desc> is incompatible with `Foo` in element 0
    - <desc> is incompatible with <non-named desc> in property `bar`
      - <desc> is incompatible with `Bar` in element 1
```

we will do

```
- in property `foo` > in element 0: <desc> is incompatible with `Foo` in element 0
    - in property `bar` > element 1: <desc> is incompatible with `Bar` in element 1
```

This approach balances between the size of the error, and the insightfulness of the error. I believe the named information is helpful, since they tend to incompatibility between something and a higher level concept, rather than just leaf node inconsistencies.

The above strategy mostly gives what we want, but it has a few issues:

1. Sometimes, it results in repeats of information, such as

```
Cannot extend `A` [1] with `B` because `B` [2] is incompatible with `A` [1] in property `foo` ...
```

To fix this, I will skip the incompatible pair if the frame is just above the root use_op that has a named upper bound.

2. Existing props collapsing optimization will hide useful named framed

e.g.
```
Cannot bla because `...` is incompatible with `...` in property `foo.bar.baz`
```

where there might be a useful frame at the `.bar` level. To fix, I make it so that the optimization will only kick in for frames without a named upper bound.

This diff only enables this for errors where there is at least one frame with named upper bound. D84853811 enables it for everything.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D84357661

fbshipit-source-id: 6c3e1fa53111796ee807da8cbec948f12ae624fa