[flow] autocomplete fix: do not suggest type imports in non-type import statements
Summary:
In
```
// a.js
export type T = number;
// b.js
import { | } from 'a';
```
we used to suggest `T` as a completion.

I'm not 100% sure this is the right direction, but it seemed a bit odd to suggest an autocomplete of a type which is an immediate error.

This diff restricts value autocompletes to things that are values.

Changelog: [fix] autocomplete on `import { | } from 'a';` will only suggest values, not types.

Reviewed By: SamChou19815

Differential Revision: D85454377

fbshipit-source-id: 9a76272926271e8d0b950d08ec2c8bea091cf843