[flow] Strip nominal identity of `InstanceT` going through `ObjKitT`
Summary:
This diff strips the nominal identity of `InstanceT` that's going through a `ObjKitT` transform, which includes operations like `Readonly`, `Partial`.

It's a long standing unsoundness issue, and it could also complicate nominal records work, if we want to reuse the `InstanceT` infra.

Changelog: [internal] Gated and default disabled for now

Reviewed By: gkz

Differential Revision: D85611050

fbshipit-source-id: ab8aa785c616c65b1e538fd73c4455ca57244d90