[flow][refactor] Introduce analysis_only mode for exhaustiveness analysis
Summary:
Later, we want to reuse the exhaustiveness analysis to compute the leftover value at midpoint, so that we can perform smart filtering for match bindings during inference.

For now, we will do the entire analysis again in the end, so it doesn't make sense to error when we are doing the analysis in the middle. It will also help prevent spurious error like

```
Cannot have a wildcard which is guarded using an `if` in the last case of a `match`
```

since the last pattern in this analysis can be a midpoint in the original pattern list.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D85465776

fbshipit-source-id: 668103390f1e0e7892459870b88c5d544fd2c766