[flow][refactor] Factor out IR of match pattern
Summary:
We generate some intermediate representation for match pattern for exhaustiveness analysis. I want to store some of these results into context, so I need to factor out the data structure part that doesn't need the context. The results will be stored in context and used in D85388235.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D85456496

fbshipit-source-id: b11683c66d580f6c2d6b3711055f693622aff8ec