[flow] Introduce special frame for custom errors
Summary:
This diff implements the core of the error calculation for custom errors.

When we see `t ~> opaque type for custom error`, we will record a special frame for this flow. Later during error IR calculation, more specific frames will be dropped. e.g. consider

```
const foo: CustomErrorT = {foo: ''};
```

Instead of mentioning that foo is incompatible with the foo prop defined in `CustomErrorT`, we will just have the error saying `object literal is incompatible with CustomErrorT` (error loc will also be at the location of the entire literal to be consistent).

For the custom error part, we will record it as an explanation. Currently, we just have a simple implementation where we just print `See relevant docs [1] (ref to doc)`. Later, we will actually fetch the comment at the `custom_error_loc` and print out relevant parts of the jsdoc (e.g. description, example)

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D86207092

fbshipit-source-id: e240a792a7c260afdcaaf0afc43741f8e509cc3f