[flow] Provide a quickfix to change match coded like a switch to more idiomatic form
Summary:
This diff implements a quickfix to change matches that are coded like a switch to a more idiomatic forms. It's attempting to target code like this

```
type DataPoint =
  | {type: 'avg', p50: null, avg: number}
  | {type: 'p50', p50: number, avg: null};

const v: number = match (dataPoint.type) {
    'avg' => dataPoint.avg,
    'p50' => dataPoint.p50,
}
```

which is written with the style of switch on disjoint union, but won't type check under match. A better way to write it is

```
type DataPoint =
  | {type: 'avg', p50: null, avg: number}
  | {type: 'p50', p50: number, avg: null};

const v: number = match (dataPoint) {
    {type:'avg', const avg} => avg,
    {type:'p50',const p50} => p50,
}
```

This diff detects this pattern by checking the following:

- whether the match root is a member expression
- whether all the cases are just simple literals
- whether there are property accesses of the form `<match_root>.prop` in case bodies.

I think the detection is probably good in 90% of cases, but just in case I missed something and the user is actually doing something totally reasonable, I decide to only offer it as a quickfix at the location of errors within match.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D85988965

fbshipit-source-id: 80b1682d9b3098a3b5a666e0167f1812d725163d