[flow][type-of-name] fix bug in finding matching result
Summary:
`flow type-of-name` searches candidates that match the target name in the export index. Before this diff, when `--exact-match-only` was passed it would stop on the first result and if the names didn't match, it would fail. With this change, it will keep looking for a match in the list of results until it gets a match.

Changelog: [internal]

Reviewed By: marcoww6

Differential Revision: D86386537

fbshipit-source-id: 5cb66af7a4dd303c8499ab312de9b52ce72b7169