[flow][MVP] Custom errors from local type aliases
Summary:
This diff implements a MVP version of custom errors. It's certainly still buggy, but functional enough for me to play it in the prod with the flag on.

In type_sig and statement.ml, we will parse the jsdoc to see whether it contains the `flowCustomError` tag. If so, we will record it as a special form of opaque type that's fully transparent for the purpose of custom errors.

Currently, there are known issues with comment attachment. e.g. the first comments will not be attached to statements but attached to the entire program. I will need to revisit it later.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D86207094

fbshipit-source-id: c497e1d7844cf3a2aa010c91bd2d3e0affe9daf4