[flow] Basic implementation to incorporate information from jsdoc for custom errors
Summary:
This diff implements the error printing infra for custom error explanations. The high level design is that the documentation should appear as an explanation for types with custom errors. The documentation is extracted from the jsdoc at the `custom_error_loc`.

This diff implements the most basic form of extraction: if available, we will take the `description` part of jsdoc as the explanation. It's certainly not production ready (we probably want to include `example` as well), but it provides the technical viability of this approach. I will use this diff as the baseline to see how custom errors behaves in real code.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D86250573

fbshipit-source-id: 894cdb84c88fe6faf07b5691b788672390a4498c