[flow][records] Initial record declaration type-checking
Summary:
Initial record declaration type-checking support, for now treating them as a class. Future diffs will add customization to this behavior.

To reduce the amount of code that needs to be maintained and be accurate, between classes and records, I have implemented part of this using the `class_of_record` and `map_record_as_class` functions.
This allows us to re-use the complex logic that exists for classes, safely.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D86821064

fbshipit-source-id: d3b13134fd634eeb43d21c4c0fa30116b4bb77b5